Move multi-agent pipeline tests to dedicated CredentialsInPipelineTest source
